Added new boss spawning logic to EnemySpawner class, updated attackspeed logic

+++ Assets/Player/Scripts/Shooting.cs
@@ -13,7 +13,7 @@ public class Shooting : MonoBehaviour
     // Update is called once per frame
     void Start()
     {
-        shootingSpeed = 2 / PlayerPrefs.GetInt("AttackSpeedLevel");
+        shootingSpeed = 2f / (float)PlayerPrefs.GetInt("AttackSpeedLevel");
         StartCoroutine(shooting());
     }
